+++ src/main/java/com/georgev22/voidchest/api/event/EventPriority.java
@@ -5,14 +5,17 @@ package com.georgev22.voidchest.api.event;
  * Listeners with lower priority will be called before those with higher priority.
  * The execution order of listeners with the same priority is undefined.
  * <p>
- * It is not recommended to make data changes in listeners with priority {@link #LOWEST}.
+ * It is not recommended to make data changes in listeners with priority {@link #LOWEST} and {@link #LOW}.
  * Data changes should be made in listeners with priority {@link #NORMAL} or {@link #HIGHEST}.
  * <p>
  * The recommended priority levels for most listeners are {@link #NORMAL} and {@link #HIGHEST}.
+ * <p>
+ * To monitor events, use {@link #MONITOR}.
  */
 public enum EventPriority {
     /**
-     * Lowest priority. Executed first.
+     * Lowest priority.
+     * Executed first.
      * Not recommended for data changes.
      */
     LOWEST(-1000),
@@ -35,10 +38,18 @@ public enum EventPriority {
     HIGH(500),
 
     /**
-     * Highest priority. Executed last.
+     * Highest priority.
      * Recommended for data changes.
      */
-    HIGHEST(1000);
+    HIGHEST(1000),
+
+    /**
+     * Monitor priority.
+     * Executed last.
+     * Not recommended for data changes but only for monitoring purposes.
+     */
+    MONITOR(2000),
+    ;
 
     private final int value;